Merge master

+++ includes/admin/submissions.php
@@ -5,24 +5,27 @@ class BuddyFormsSubmissionPage {
 	 * @var BuddyForms_Submissions_List_Table
 	 */
 	private $bf_submissions_table;
+	/*
+	 * @var default capability
+	 */
+	private $bf_submission_capability = 'read';
 
 	public function __construct() {
 		add_action( 'admin_menu', array( $this, 'buddyforms_create_submissions_page' ) );
 
 		add_filter( 'set-screen-option', array( $this, 'buddyforms_submissions_set_option' ), 10, 3 );
 		add_action( 'admin_init', array( $this, 'redirect_after_delete' ) );
-
 	}
 
 	public function buddyforms_create_submissions_page() {
-		$buddyforms_submission_admin_page = add_submenu_page( 'edit.php?post_type=buddyforms', __( 'Submissions', 'buddyforms' ), __( 'Submissions', 'buddyforms' ), 'activate_plugins', 'buddyforms_submissions', array( $this, 'buddyforms_submissions_screen' ) );
+		$buddyforms_submission_admin_page = add_submenu_page( 'edit.php?post_type=buddyforms', __( 'Submissions', 'buddyforms' ), __( 'Submissions', 'buddyforms' ), $this->bf_submission_capability, 'buddyforms_submissions', array( $this, 'buddyforms_submissions_screen' ) );
 		add_action( "load-$buddyforms_submission_admin_page", array( $this, 'buddyforms_submissions_add_options' ) );
 	}
 
 	public function buddyforms_submissions_add_options() {
 		$option = 'per_page';
 		$args   = array(
-			'label'   => 'Entries',
+			'label'   => __( 'Entries', 'buddyforms' ),
 			'default' => 10,
 			'option'  => 'entries_per_page'
 		);
@@ -35,72 +38,98 @@ class BuddyFormsSubmissionPage {
 		global $buddyforms, $current_screen, $parent_file, $form_slug, $post_id;
 
 		// Check that the user is allowed to update options
-		if ( ! current_user_can( 'manage_options' ) ) {
+		if ( ! current_user_can( $this->bf_submission_capability ) ) {
 			wp_die( __( 'You do not have sufficient permissions to access this page.', 'buddyforms' ) );
-		} ?>
+		}
+		?>
         <div id="post" class="bf_admin_wrap wrap">
-			<?php
-			include( BUDDYFORMS_INCLUDES_PATH . '/admin/admin-header.php' );
-			?>
-            <hr style="margin-bottom: 0px !important;"/><?php
-			$this->bf_submissions_table->prepare_items();
-			if ( isset( $_GET['form_slug'] ) ) {
-				$current_screen->set_parentage( $parent_file );
-				//$current_screen->render_screen_options();
-				$current_screen->render_screen_meta();
-			}
-
-			?>
-
-            <div id="icon-users" class="icon32"><br/></div>
-            <div id="buddyforms_admin_main_menu" class="">
-                <ul>
-                    <li>
-
-                        <h4>Select a form to display the submissions</h4>
-                        <script type="text/javascript">
-                          jQuery(document).ready(function (jQuery) {
-                            jQuery('#buddyforms_admin_menu_submissions_form_select').change(function () {
-                              window.location = '?post_type=buddyforms&page=buddyforms_submissions&form_slug=' + this.value
-                            })
-
-                            jQuery('.metabox-prefs input:checkbox').each(function () {
-                              var colID = jQuery(this).attr('id')
-                              var hasCheckedAtt = document.getElementById(colID).hasAttribute('checked')
-                              if (!hasCheckedAtt) {
-                                document.getElementById(colID).checked = false
-                              }
-
-                            })
-
-                          })
-                        </script>
-                        <select id="buddyforms_admin_menu_submissions_form_select">
-                            <option value="none">Select Form</option>
-							<?php foreach ( $buddyforms as $form_slug => $form ) { ?>
-                                <option <?php isset( $_GET['form_slug'] ) ? selected( $_GET['form_slug'], $form_slug ) : ''; ?> value="<?php echo $form_slug ?>">
-									<?php echo $form['name']; ?>
-                                </option>
-							<?php } ?>
-                        </select>
-                    </li>
-                </ul>
-            </div>
-
-			<?php if ( isset( $_GET['form_slug'] ) && ! isset( $_GET['entry'] ) ) { ?>
-                <form id="filter" method="get">
-                    <input type="hidden" name="page" value="<?php echo $_REQUEST['page'] ?>"/>
-					<?php $this->bf_submissions_table->display(); ?>
-                </form>
-			<?php } ?>
-
-			<?php if ( isset( $_GET['action'] ) && isset( $_GET['entry'] ) ) {
-				$form_slug = get_post_meta( $_GET['entry'], '_bf_form_slug', true );
-				$post_id   = $_GET['entry'];
-				require_once( BUDDYFORMS_INCLUDES_PATH . 'admin/submission-single.php' );
-			} ?>
-        </div>
 		<?php
+		include( BUDDYFORMS_INCLUDES_PATH . '/admin/admin-header.php' );
+		?>
+        <hr style="margin-bottom: 0px !important;"/><?php
+		$this->bf_submissions_table->prepare_items();
+		$selected_form = '';
+		if ( isset( $_GET['form_slug'] ) ) {
+			$current_screen->set_parentage( $parent_file );
+			$current_screen->render_screen_meta();
+			$selected_form = $_GET['form_slug'];
+		}
+
+		?>
+
+        <div id="icon-users" class="icon32"><br/></div>
+        <div id="buddyforms_admin_main_menu" class="">
+            <ul>
+                <li>
+                    <h4><?php _e( 'Select a form to display the submissions', 'buddyforms' ) ?></h4>
+                    <script type="text/javascript">
+                        jQuery(document).ready(function (jQuery) {
+                            jQuery('#buddyforms_admin_menu_submissions_form_select').change(function () {
+                                window.location = '?post_type=buddyforms&page=buddyforms_submissions&form_slug=' + this.value
+                            });
+                            jQuery('.metabox-prefs input:checkbox').each(function () {
+                                var colID = jQuery(this).attr('id');
+                                var hasCheckedAtt = document.getElementById(colID).hasAttribute('checked');
+                                if (!hasCheckedAtt) {
+                                    document.getElementById(colID).checked = false
+                                }
+                            })
+                        })
+                    </script>
+                    <select id="buddyforms_admin_menu_submissions_form_select">
+                        <option value="none"><?php _e( 'Select Form', 'buddyforms' ) ?></option>
+						<?php foreach ( $buddyforms as $form_slug => $form ) : ?>
+							<?php if ( ! $this->has_the_capability( $form_slug ) ) : ?>
+								<?php continue; ?>
+							<?php endif; ?>
+                            <option <?php selected( $selected_form, $form_slug ) ?> value="<?php echo $form_slug ?>"><?php echo $form['name']; ?></option>
+						<?php endforeach; ?>
+                    </select>
+                </li>
+            </ul>
+        </div>
+
+		<?php if ( isset( $_GET['form_slug'] ) ) : ?>
+			<?php if ( $this->has_the_capability( $_GET['form_slug'] ) ) : ?>
+				<?php if ( ! isset( $_GET['entry'] ) ) { ?>
+                    <form id="filter" method="get">
+                        <input type="hidden" name="page" value="<?php echo $_REQUEST['page'] ?>"/>
+						<?php $this->bf_submissions_table->display(); ?>
+                    </form>
+				<?php } ?>
+
+				<?php if ( isset( $_GET['action'] ) && isset( $_GET['entry'] ) ) {
+					$form_slug = get_post_meta( $_GET['entry'], '_bf_form_slug', true );
+					$post_id   = $_GET['entry'];
+					require_once( BUDDYFORMS_INCLUDES_PATH . 'admin/submission-single.php' );
+				} ?>
+                </div>
+			<?php else: ?>
+                <strong><?php _e( 'You do not have sufficient permissions to access this page.', 'buddyforms' ) ?></strong>
+			<?php endif; ?>
+		<?php endif; ?>
+		<?php
+	}
+
+	/**
+	 * Determine if the user have the capability to get the submission for this form.
+	 * Note: This return true for user with admin role, checking the `activate_plugins` capability
+	 *
+	 * @since 2.3.1
+	 *
+	 * @param $form_slug
+	 *
+	 * @return bool
+	 */
+	function has_the_capability( $form_slug ) {
+		if ( empty( $form_slug ) ) {
+			return false;
+		}
+		if ( current_user_can( 'activate_plugins' ) ) {
+			return true;
+		}
+
+		return current_user_can( 'buddyforms_' . $form_slug . '_admin-submission' );
 	}
 
 	function redirect_after_delete() {
@@ -109,7 +138,7 @@ class BuddyFormsSubmissionPage {
 		$action    = isset( $_GET['action'] ) ? $_GET['action'] : "";
 		$entry     = isset( $_GET['post'] ) ? $_GET['post'] : "";
 		$form_slug = isset( $_GET['form_slug'] ) ? $_GET['form_slug'] : "";
-		if ( $action === 'delete' ) {
+		if ( $action === 'delete' && $this->has_the_capability( $form_slug ) ) {
 			$buddyFData = isset( $buddyforms[ $form_slug ]['form_fields'] ) ? $buddyforms[ $form_slug ]['form_fields'] : [];
 			foreach ( $buddyFData as $key => $value ) {
 
@@ -120,13 +149,14 @@ class BuddyFormsSubmissionPage {
 					$can_delete_files = isset( $value['delete_files'] ) ? true : false;
 					if ( $can_delete_files ) {
 						// If true then Delete the files attached to the entry
-						$column_val   = get_post_meta( $entry, $field, true );
-						$attachmet_id = explode( ",", $column_val );
-						foreach ( $attachmet_id as $id ) {
-							wp_delete_attachment( $id, true );
+						$column_val = get_post_meta( $entry, $field, true );
+						if ( ! empty( $column_val ) ) {
+							$attachmet_id = explode( ",", $column_val );
+							foreach ( $attachmet_id as $id ) {
+								wp_delete_attachment( $id, true );
+							}
 						}
 					}
-
 				}
 			}
 		}
@@ -187,13 +217,14 @@ class BuddyForms_Submissions_List_Table extends WP_List_Table {
 	function column_ID( $item ) {
 		global $buddyforms;
 
-		$actions = array(
-			'edit'   => sprintf( '<a href="post.php?post=%s&action=%s">Edit</a>', $item->ID, 'edit' ),
-			'delete' => '<a href="' . get_delete_post_link( $item->ID, '', true ) . '" class="submitdelete deletion" onclick="return confirm(\'Are you sure you want to delete that entry?\');" title="Delete">Delete</a>',
+		$form_slug = isset( $_GET['form_slug'] ) ? $_GET['form_slug'] : "";
+		$actions   = array(
+			'edit'   => sprintf( '<a href="post.php?post=%s&action=%s">%s</a>', $item->ID, 'edit', __( 'Edit', 'buddyforms' ) ),
+			'delete' => '<a href="' . get_delete_post_link( $item->ID, '', true ) . '&form_slug=' . $form_slug . '" class="submitdelete deletion" onclick="return confirm(\'' . __( 'Are you sure you want to delete that entry?', 'buddyforms' ) . '\');" title="' . __( 'Delete', 'buddyforms' ) . '">' . __( 'Delete', 'buddyforms' ) . '</a>',
 		);
 
 		if ( isset( $buddyforms[ $_GET['form_slug'] ]['post_type'] ) && $buddyforms[ $_GET['form_slug'] ]['post_type'] == 'bf_submissions' ) {
-			$actions['edit'] = sprintf( '<a href="?post_type=buddyforms&page=%s&action=%s&entry=%s&form_slug=%s">View Submission</a>', $_REQUEST['page'], 'edit', $item->ID, $_GET['form_slug'] );
+			$actions['edit'] = sprintf( '<a href="?post_type=buddyforms&page=%s&action=%s&entry=%s&form_slug=%s">%s</a>', $_REQUEST['page'], 'edit', $item->ID, $_GET['form_slug'], __( 'View Submission', 'buddyforms' ) );
 		}
 
 		// Return the title contents
@@ -215,6 +246,14 @@ class BuddyForms_Submissions_List_Table extends WP_List_Table {
 		}
 		if ( $column_name == 'Date' ) {
 			echo get_the_date( 'F j, Y', $item->ID );
+		}
+		if ( $column_name == 'Author' ) {
+			$post = get_post( $item->ID );
+			if ( ! empty( $post->post_author ) ) {
+				echo apply_filters( 'bf_submission_column_default_author_meta', get_the_author_meta( 'nickname', $post->post_author ), $post->post_author );
+			} else {
+				_e( 'Anonymous', 'buddyforms' );
+			}
 		}
 	}
 
@@ -273,6 +312,15 @@ class BuddyForms_Submissions_List_Table extends WP_List_Table {
 				break;
 			case 'status':
 				$bf_value = buddyforms_get_post_status_readable( get_post_status( $item->ID ) );
+				break;
+			case 'user_login':
+				$author_id = ( ! empty( $item->post_author ) ) ? $item->post_author : 0;
+				if ( ! empty( $author_id ) ) {
+					$author = get_user_by( 'ID', $author_id );
+					if ( $author instanceof WP_User ) {
+						$bf_value = $author->user_login;
+					}
+				}
 				break;
 			default:
 				if ( is_array( $bf_value ) ) {
@@ -333,13 +381,14 @@ class BuddyForms_Submissions_List_Table extends WP_List_Table {
 		global $buddyforms;
 
 		$columns = array(
-			'ID'   => 'ID',
-			'Date' => 'Date',
+			'ID'     => 'ID',
+			'Author' => __( 'Author', 'buddyforms' ),
+			'Date'   => __( 'Date', 'buddyforms' ),
 		);
 
 		if ( isset( $_GET['form_slug'] ) && isset( $buddyforms[ $_GET['form_slug'] ]['form_fields'] ) ) {
 			foreach ( $buddyforms[ $_GET['form_slug'] ]['form_fields'] as $key => $field ) {
-				if ( $field['slug'] != 'user_pass' ) {
+				if ( ! empty( $field['slug'] ) && ! empty( $field['name'] ) && $field['slug'] != 'user_pass' ) {
 					$columns[ $field['slug'] ] = $field['name'];
 				}
 			}
